Add record converter implementation

+++ app/src/main/java/com/tomclaw/nimpas/journal/Record.kt
@@ -3,13 +3,13 @@ package com.tomclaw.nimpas.journal
 interface Record
 
 data class Group(
-        val id: Int,
+        val id: Long,
         val title: String,
         val records: List<Record>
 ) : Record
 
 data class Password(
-        val id: Int,
+        val id: Long,
         val title: String,
         val username: String?,
         val password: String?,
@@ -18,7 +18,7 @@ data class Password(
 ) : Record
 
 data class Card(
-        val id: Int,
+        val id: Long,
         val title: String,
         val number: String,
         val expiration: String?,
@@ -27,7 +27,7 @@ data class Card(
 ) : Record
 
 data class Note(
-        val id: Int,
+        val id: Long,
         val title: String,
         val text: String
 ) : Record
